# docs/swagger.yaml
basePath: /api/v2
definitions:
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse:
    properties:
      code:
        type: integer
      details:
        type: object
      message:
        type: string
    type: object
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarCreateRequest:
    properties:
      cityIbge:
        type: string
      isActive:
        type: boolean
      name:
        type: string
      scope:
        type: string
      state:
        type: string
      timezone:
        example: America/Sao_Paulo
        type: string
    required:
    - name
    - scope
    - timezone
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDateCreateRequest:
    properties:
      calendarId:
        type: integer
      holidayDate:
        type: string
      label:
        type: string
      recurrent:
        type: boolean
    required:
    - calendarId
    - holidayDate
    - label
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDateDeleteRequest:
    properties:
      id:
        type: integer
    required:
    - id
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDateResponse:
    properties:
      calendarId:
        type: integer
      holidayDate:
        type: string
      id:
        type: integer
      label:
        type: string
      recurrent:
        type: boolean
      timezone:
        type: string
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDatesListResponse:
    properties:
      dates:
        items:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDateResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.PaginationResponse'
    type: object
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarResponse:
    properties:
      cityIbge:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      scope:
        type: string
      state:
        type: string
      timezone:
        type: string
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarUpdateRequest:
    properties:
      cityIbge:
        type: string
      id:
        type: integer
      isActive:
        type: boolean
      name:
        type: string
      scope:
        type: string
      state:
        type: string
      timezone:
        example: America/Sao_Paulo
        type: string
    required:
    - id
    - name
    - scope
    - timezone
    type: object
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarsListResponse:
    properties:
      calendars:
        items:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarResponse'
        type: array
      pagination:
        $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.PaginationResponse'
    type: object
...
  github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.PaginationResponse:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
host: api.gca.dev.br
info:
  contact:
    email: projeto.toq@gmail.com
    name: TOQ Development Team
    url: https://github.com/projeto-toq/toq_server
  description: TOQ Server - Real Estate HTTP API Server with hexagonal architecture.
  license:
    name: MIT
    url: https://github.com/projeto-toq/toq_server/blob/main/LICENSE
  termsOfService: http://swagger.io/terms/
  title: TOQ Server API
  version: "2.0"
paths:
...
  /admin/holidays/calendars:
    get:
      consumes:
      - application/json
      description: Returns holiday calendars with filtering and pagination options.
      parameters:
      - description: Calendar scope (NATIONAL|STATE|CITY)
        example: '"STATE"'
        in: query
        name: scope
        type: string
      - description: State abbreviation
        example: '"SP"'
        in: query
        name: state
        type: string
      - description: City IBGE code
        example: '"3509502"'
        in: query
        name: cityIbge
        type: string
      - description: Free text search
        example: '"Christmas"'
        in: query
        name: search
        type: string
      - description: Filter active calendars
        example: true
        in: query
        name: onlyActive
        type: boolean
      - description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - description: Page size
        example: 20
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List holiday calendars
      tags:
      - Admin Holidays
    post:
      consumes:
      - application/json
      description: Creates a new holiday calendar available for scheduling operations.
      parameters:
      - description: Calendar payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarCreateRequest'
        x-example: '{"name":"Feriados Sao Paulo","scope":"STATE","state":"SP","cityIbge":"","isActive":true,"timezone":"America/Sao_Paulo"}'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create holiday calendar
      tags:
      - Admin Holidays
    put:
      consumes:
      - application/json
      description: Updates an existing holiday calendar metadata.
      parameters:
      - description: Calendar update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarUpdateRequest'
        x-example: '{"id":42,"name":"Calendario Campinas","scope":"CITY","state":"SP","cityIbge":"3509502","isActive":true,"timezone":"America/Sao_Paulo"}'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update holiday calendar
      tags:
      - Admin Holidays
...
  /admin/holidays/dates:
    delete:
      consumes:
      - application/json
      description: Removes a holiday date from a calendar.
      parameters:
      - description: Holiday date delete request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDateDeleteRequest'
        x-example: '{"id":1001}'
      produces:
      - application/json
      responses:
        "204":
          description: Holiday date deleted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete holiday date
      tags:
      - Admin Holidays
    get:
      consumes:
      - application/json
      description: Returns holiday dates for a calendar with optional range filters.
      parameters:
      - description: Calendar identifier
        example: 42
        in: query
        name: calendarId
        required: true
        type: integer
      - description: Start date (RFC3339)
        example: '"2025-12-01T00:00:00Z"'
        in: query
        name: from
        type: string
      - description: End date (RFC3339)
        example: '"2026-01-10T23:59:59Z"'
        in: query
        name: to
        type: string
      - description: Page number
        example: 1
        in: query
        name: page
        type: integer
      - description: Page size
        example: 50
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDatesListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List holiday dates
      tags:
      - Admin Holidays
    post:
      consumes:
      - application/json
      description: Registers a new holiday date within a calendar.
      parameters:
      - description: Holiday date payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDateCreateRequest'
        x-example: '{"calendarId":42,"holidayDate":"2025-12-25T00:00:00Z","label":"Christmas","recurrent":true}'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.HolidayCalendarDateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_projeto-toq_toq_server_internal_adapter_left_http_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create holiday date
      tags:
      - Admin Holidays
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"